Adding modified README, fixing lint errors
