+++ forte/data/ontology/__init__.py
@@ -1,2 +1,3 @@
+from forte.data.ontology.ontology_code_generator import *
 from forte.data.ontology.top import *
 from forte.data.ontology.core import *
