+++ setup.py
@@ -12,8 +12,7 @@ setuptools.setup(
     description="NLP pipeline",
     long_description=long_description,
     license='Apache License Version 2.0',
-
-    packages=setuptools.find_packages(include=['forte.*', 'ft.*']),
+    packages=setuptools.find_packages(include=['forte.*', 'ft.*', 'generate_ontology.*']),
     platforms='any',
 
     install_requires=[
@@ -36,6 +35,11 @@ setuptools.setup(
     },
     package_data={
     },
+    entry_points={
+          'console_scripts': [
+              'generate_ontology = generate_ontology.__main__:main'
+          ]
+      },
     classifiers=[
         'Intended Audience :: Developers',
         'Intended Audience :: Education',
